backend/requirements.txt
fastapi
uvicorn[standard]
websockets
google-genai
python-dotenv
PyGithub
